pyproject.toml
[tool.poetry]
name = "opendata-ph-data-platform"
version = "0.1.0"
description = ""
authors = ["aldencabajar <aldenacabajar@gmail.com>"]
license = "Apache"
readme = "README.md"
packages = [
    { include = "opendata_ph", from="src"}
]


[tool.poetry.dependencies]
python = ">=3.10,<3.13"
playwright = "^1.53.0"
openpyxl = "^3.1.5"
pandas = "^2.3.1"
apache-airflow = "^3.0.3"
pyarrow = "^21.0.0"
fastparquet = "^2024.11.0"
apache-airflow-providers-standard = "^1.4.1"
duckdb = "^1.3.2"
dbt-core = "^1.10.6"
dbt-duckdb = "^1.9.4"
bs4 = "^0.0.2"
pydantic = "^2.11.7"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

